Rehaul group visualization
This commit expands the network visualization from two
design groups to a maximum of eight groups. Nodes are
color-coded by the group in which the species/OTU/oligotype
is most abundant. The length of the colorblindness-accessible
color palette limits the unique groups to eight. This commit
also adds a legend and image size scaling.